+++ punchbowl/level1/tests/test_sqrt.py
@@ -1,4 +1,4 @@
-from datetime import datetime
+from datetime import UTC, datetime
 
 import numpy as np
 import pytest