+++ punchbowl/level2/flow.py
@@ -1,5 +1,5 @@
 from glob import glob
-from datetime import datetime
+from datetime import UTC, datetime
 
 import numpy as np
 from astropy.nddata import StdDevUncertainty
@@ -65,7 +65,7 @@ def level2_core_flow(data_list: list[str] | list[NDCube],
                      for cube, this_voter_filenames in zip(data_list, ordered_voters, strict=True)]
         output_data = merge_many_polarized_task(data_list, trefoil_wcs)
     else:
-        output_dateobs = datetime.now().isoformat()
+        output_dateobs = datetime.now(UTC).isoformat()
         output_datebeg = output_dateobs
         output_dateend = output_datebeg
 
@@ -76,7 +76,7 @@ def level2_core_flow(data_list: list[str] | list[NDCube],
         meta=NormalizedMetadata.load_template("PTM", "2"),
     )
 
-    output_data.meta["DATE"] = datetime.now().isoformat()
+    output_data.meta["DATE"] = datetime.now(UTC).isoformat()
     output_data.meta["DATE-AVG"] = output_dateobs
     output_data.meta["DATE-OBS"] = output_dateobs
     output_data.meta["DATE-BEG"] = output_datebeg