+++ punchbowl/level3/polarization.py
@@ -1,4 +1,4 @@
-from datetime import datetime
+from datetime import UTC, datetime
 
 import astropy.units as u
 import numpy as np
@@ -21,9 +21,9 @@ def convert_polarization(
                             NDCube(data=input_data[i].data,
                                    wcs=input_data.wcs.dropaxis(2),
                                    meta={"POLAR": angle}))
-                       for (label, i, angle) in zip(["M", "Z", "P"],
-                                                    [0, 1, 2],
-                                                    [-60*u.deg, 0*u.deg, 60*u.deg], strict=False)]
+                           for (label, i, angle) in zip(["M", "Z", "P"],
+                                                        [0, 1, 2],
+                                                        [-60 * u.deg, 0 * u.deg, 60 * u.deg], strict=False)]
     data_collection = NDCollection(collection_contents, aligned_axes="all")
 
     resolved_data_collection = resolve(data_collection, "BpB", imax_effect=False)
@@ -32,7 +32,7 @@ def convert_polarization(
     new_wcs = input_data.wcs.copy()
 
     output_meta = NormalizedMetadata.load_template("PTM", "3")
-    output_meta["DATE"] = datetime.now().isoformat()
+    output_meta["DATE"] = datetime.now(UTC).isoformat()
     output_meta["DATE-AVG"] = input_data.meta["DATE-AVG"].value
     output_meta["DATE-OBS"] = input_data.meta["DATE-OBS"].value
     output_meta["DATE-BEG"] = input_data.meta["DATE-BEG"].value