+++ punchbowl/level3/stellar.py
@@ -1,18 +1,96 @@
 from math import floor
-from datetime import datetime
+from datetime import UTC, datetime
 
+import astropy.units as u
 import numpy as np
 import remove_starfield
 from astropy.wcs import WCS
 from dateutil.parser import parse as parse_datetime_str
-from ndcube import NDCube
+from ndcube import NDCollection, NDCube
 from prefect import flow, get_run_logger
 from remove_starfield import ImageHolder, ImageProcessor, Starfield
 from remove_starfield.reducers import PercentileReducer
+from solpolpy import resolve
 
 from punchbowl.data import NormalizedMetadata, load_ndcube_from_fits
-from punchbowl.data.wcs import calculate_helio_wcs_from_celestial
+from punchbowl.data.wcs import calculate_helio_wcs_from_celestial, get_p_angle
 from punchbowl.prefect import punch_task
+
+
+def to_celestial(input_data: NDCube) -> NDCube:
+    """
+    Convert polarization from mzpsolar to Celestial frame.
+
+    All images need their polarization converted to Celestial frame
+    to generate the background starfield model.
+    """
+    # Create a data collection for M, Z, P components
+    mzp_angles = [-60, 0, 60]*u.degree
+
+    # Compute new angles for celestial frame
+    cel_north_offset = get_p_angle(time=input_data[0].meta["DATE-OBS"].value)
+    new_angles = mzp_angles - cel_north_offset
+
+    collection_contents = [
+        (label,
+         NDCube(data=input_data[i].data,
+                wcs=input_data.wcs.dropaxis(2),
+                meta={"POLAR": angle}))
+        for label, i, angle in zip(["M", "Z", "P"], [0, 1, 2], mzp_angles, strict=False)
+    ]
+    data_collection = NDCollection(collection_contents, aligned_axes="all")
+
+    # Resolve data to celestial frame
+    celestial_data_collection = resolve(data_collection, "npol", out_angles=new_angles, imax_effect=False)
+
+    valid_keys = [key for key in celestial_data_collection if key != "alpha"]
+    new_data = [celestial_data_collection[key].data for key in valid_keys]
+    new_wcs = input_data.wcs.copy()
+
+    output_meta = NormalizedMetadata.load_template("PTM", "3")
+    output_meta["DATE-OBS"] = input_data.meta["DATE-OBS"].value
+
+    output = NDCube(data=new_data, wcs=new_wcs, meta=output_meta)
+    output.meta.history.add_now("LEVEL3-convert2celestial", "Convert mzpsolar to Celestial")
+
+    return output
+
+
+def from_celestial(input_data: NDCube) -> NDCube:
+    """
+    Convert polarization from Celestial frame to mzpsolar.
+
+    All images need their polarization converted back to Solar frame
+    after removing the stellar polarization.
+    """
+    # Create a data collection for M, Z, P components
+    mzp_angles = [-60, 0, 60]*u.degree
+    # Compute new angles for celestial frame
+    cel_north_offset = get_p_angle(time=input_data[0].meta["DATE-OBS"].value)
+    new_angles = mzp_angles - cel_north_offset
+    collection_contents = [
+        (f"{angle.value} deg",
+         NDCube(data=input_data[i].data,
+                wcs=input_data.wcs.dropaxis(2),
+                meta={"POLAR": angle}))
+        for i, angle in enumerate(new_angles)
+    ]
+    data_collection = NDCollection(collection_contents, aligned_axes="all")
+
+    # Resolve data to mzpsolar frame
+    solar_data_collection = resolve(data_collection, "mzpsolar", imax_effect=False)
+
+    valid_keys = [key for key in solar_data_collection if key != "alpha"]
+    new_data = [solar_data_collection[key].data for key in valid_keys]
+    new_wcs = input_data.wcs.copy()
+
+    output_meta = NormalizedMetadata.load_template("PTM", "2")
+    output_meta["DATE-OBS"] = input_data.meta["DATE-OBS"].value
+
+    output = NDCube(data=new_data, wcs=new_wcs, meta=output_meta, uncertainty=input_data.uncertainty)
+    output.meta.history.add_now("LEVEL3-convert2mzpsolar", "Convert Celestial to mzpsolar")
+
+    return output
 
 
 class PUNCHImageProcessor(ImageProcessor):
@@ -27,6 +105,7 @@ class PUNCHImageProcessor(ImageProcessor):
     def load_image(self, filename: str) -> ImageHolder:
         """Load an image."""
         cube = load_ndcube_from_fits(filename, key=self.key)
+        cube = to_celestial(cube)
         data = cube.data[self.layer]
         if self.apply_mask:
             data[np.isclose(cube.uncertainty.array[self.layer], 0, atol=1E-30)] = np.nan
@@ -44,7 +123,7 @@ def generate_starfield_background(
     logger = get_run_logger()
 
     if reference_time is None:
-        reference_time = datetime.now()
+        reference_time = datetime.now(UTC)
     elif isinstance(reference_time, str):
         reference_time = parse_datetime_str(reference_time)
 
@@ -113,7 +192,7 @@ def generate_starfield_background(
     meta["DATE-BEG"] = reference_time.isoformat()
     meta["DATE-END"] = reference_time.isoformat()
     meta["DATE-AVG"] = reference_time.isoformat()
-    meta["DATE"] = datetime.now().isoformat()
+    meta["DATE"] = datetime.now(UTC).isoformat()
 
     out_wcs, _ = calculate_helio_wcs_from_celestial(starfield_m.wcs, meta.astropy_time, starfield_m.starfield.shape)
     output = NDCube(np.stack([starfield_m.starfield, starfield_z.starfield, starfield_p.starfield], axis=0),
@@ -159,20 +238,8 @@ def subtract_starfield_background_task(data_object: NDCube,
         # data_wcs = calculate_celestial_wcs_from_helio(data_object.wcs.celestial,
         #                                               data_object.meta.astropy_time,
         #                                               data_object.data.shape[-2:])
-        map_scale = 0.01
-        shape = [floor(132 / map_scale), floor(360 / map_scale)]
-        starfield_wcs = WCS(naxis=2)
-        # n.b. it seems the RA wrap point is chosen so there's 180 degrees
-        # included on either side of crpix
-        crpix = [shape[1] / 2 + .5, shape[0] / 2 + .5]
-        starfield_wcs.wcs.crpix = crpix
-        starfield_wcs.wcs.crval = 270, -23.5
-        starfield_wcs.wcs.cdelt = map_scale, map_scale
-        starfield_wcs.wcs.ctype = "RA---CAR", "DEC--CAR"
-        starfield_wcs.wcs.cunit = "deg", "deg"
-        starfield_wcs.array_shape = shape
 
-        starfield_model = Starfield(np.stack((star_datacube.data, star_datacube.uncertainty.array)), starfield_wcs)
+        starfield_model = Starfield(np.stack((star_datacube.data, star_datacube.uncertainty.array)), star_datacube.wcs)
 
         subtracted = starfield_model.subtract_from_image(
             NDCube(data=np.stack((data_object.data, data_object.uncertainty.array)),
@@ -183,9 +250,8 @@ def subtract_starfield_background_task(data_object: NDCube,
         data_object.data[...] = subtracted.subtracted[0]
         data_object.uncertainty.array[...] -= subtracted.subtracted[1]
         data_object.meta.history.add_now("LEVEL3-subtract_starfield_background", "subtracted starfield background")
-        output = data_object
+        output = from_celestial(data_object)
     logger.info("subtract_starfield_background finished")
-
 
     return output